WIP: add CAPI chart
this chart deploys a capi cluster onto the cloud.
this is an attempt to create an easier way for users to create an STFC Cloud compliant k8s cluster.

With this umbrella chart, we can add logging and security features as subcharts to keep these clusters complaint.

it also saves us from having to maintain a separate cloud-capi-values repo.

It also allows us to test and version changes better